requirements.txt
ffmpeg-python==0.2.0
future==1.0.0
greenlet==3.3.1
playwright==1.58.0
pyee==13.0.1
typing_extensions==4.15.0
